pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal-share"
version = "1.0.0"
description = "Shared pwsh session wrapper exposing an MCP server for collaborative terminal use."
requires-python = ">=3.11"
dependencies = [
    "pywinpty>=2.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[tool.setuptools.packages.find]
include = ["terminal_share*"]
exclude = ["tests*"]
